# api/docs/swagger.yaml
definitions:
...
  api.ErrResponse:
    properties:
      details:
        items:
          type: string
        type: array
      error:
        type: string
    type: object
...
  api.verifyEmailOTPBody:
    properties:
      otp:
        type: string
    required:
    - otp
    type: object
...
info:
  contact: {}
  description: modDNS REST API
  title: modDNS REST API
  version: "1.0"
paths:
...
  /api/v1/verify/email/otp/confirm:
    post:
      consumes:
      - application/json
      description: Verifies the 6-digit OTP provided by the authenticated user
      parameters:
      - description: OTP verification request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.verifyEmailOTPBody'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      summary: Confirm email verification OTP
      tags:
      - Verification
  /api/v1/verify/email/otp/request:
    post:
      description: Generates and sends a 6-digit OTP to verify the authenticated user's
        email
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      summary: Request email verification OTP
      tags:
      - Verification
...
swagger: "2.0"